requirements.txt
tweepy==4.14.0
google-generativeai==0.8.3
APScheduler==3.10.4
python-dotenv==1.0.1
streamlit>=1.34.0
requests>=2.31.0
